// testing/types/timekeeper.d.ts
export interface components {
    schemas: {
// ...
        InsertablePackagedTimeSplit: {
            time_split: components["schemas"]["InsertableTimeSplit"];
            timers: components["schemas"]["InsertableTimeSplitTimer"][];
        };
// ...
        InsertableTimer: {
            end_time: components["schemas"]["Timestamptz"];
            start_time: components["schemas"]["Timestamptz"];
            tags: number[];
            /** Format: int32 */
            time_split_timer: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}